+++ history.md
@@ -1,3 +1,8 @@
+# 2023-01-30
+
+- Always show `cancellable` progress as notification without check
+  `notification.statusLineProgress`.
+
 # 2023-01-29
 
 - Exclude `source` actions when request code actions with range.